requirements.txt
alembic==1.13.2
blinker==1.8.2
CacheControl==0.14.0
cachelib==0.13.0
cachetools==5.3.3
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
ecdsa==0.19.0
email_validator==1.1.3
fastapi-sessions==0.3.2
firebase-admin==6.5.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gcloud==0.18.3
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
hpack==4.0.0
httplib2==0.22.0
httptools==0.6.1
jwcrypto==1.5.6
Mako==1.3.5
msgpack==1.0.8
msgspec==0.18.6
oauth2client==4.1.3
passlib==1.7.4
paypalrestsdk==1.13.3
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
pyparsing==3.1.2
Pyrebase4==4.8.0
python-dotenv==1.0.1
python-jose==3.3.0
python-jwt==4.1.0
PyYAML==6.0.1
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
setuptools==70.0.0
six==1.16.0
SQLAlchemy==2.0.30
typer==0.12.3
typer-slim==0.12.3
uritemplate==4.1.1
urllib3==1.26.18
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
WTForms==3.1.2
